+++ examples/exit_after_n_jobs_get_successfully_processed/main.go
@@ -6,6 +6,7 @@
 package main
 
 import (
+	"fmt"
 	"log"
 	"time"
 
@@ -13,14 +14,27 @@ import (
 )
 
 func main() {
-	// total workers
-	totalWorkers := 10
-	// max number of pending jobs
-	maxNumberPendingJobs := 15
-	// do not show messages about the pool processing
-	verbose := false
+	var (
+		// total workers
+		totalWorkers uint = 10
+		// max number of pending jobs
+		maxNumberPendingJobs uint = 150
+		// do not show messages about the pool processing
+		verbose = false
+	)
 
-	pool := goworkerpool.NewPool(totalWorkers, maxNumberPendingJobs, verbose)
+	pool, err := goworkerpool.NewPoolWithOptions(goworkerpool.PoolOptions{
+		TotalInitialWorkers:          totalWorkers,
+		MaxOperationsInQueue:         maxNumberPendingJobs,
+		MaxWorkers:                   20,
+		WaitUntilInitialWorkersAreUp: true,
+		LogVerbose:                   verbose,
+	})
+
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
 
 	// add the worker handler function
 	pool.SetWorkerFunc(func(data interface{}) bool {
@@ -33,9 +47,6 @@ func main() {
 		return true
 	})
 
-	// start up the workers and wait until them are up
-	pool.StartWorkersAndWait()
-
 	// enqueue jobs in a separate goroutine
 	go func() {
 		for i := 0; i < 100; i++ {
@@ -43,6 +54,6 @@ func main() {
 		}
 	}()
 
-	// wait until 30 jobs get successfully processed
+	// wait until at least 30 jobs get successfully processed
 	pool.WaitUntilNSuccesses(30)
 }