+++ strings.go
@@ -1,0 +1,9 @@
+// ** goworkerpool.com  **********************************************************************************************
+// ** github.com/enriquebris/goworkerpool  																			**
+// ** v0.10.0  *******************************************************************************************************
+
+package goworkerpool
+
+const (
+	messageKillAllWorkersInProgress = "KillAllWorkers() operation in progress"
+)